feat: add AddExpensePage and TopNavigation component, integrate quick expense capture in DashboardPage, and enhance styles for navigation links
